Adds reprotest for .deb files
More pytest-based reprotest invocations, this time focusing on .deb
files. Replaces the CircleCI repro tests that manually compared hashes.

Also modifies build script to support commit hash

The logic assumed we were always building from a prod release tag.
As a result, the CI logic was reimplementing the tarball mangling.
Let's make the script more flexible, so we can run the script in CI and
thereby get a bit more test coverage for it.

Modifies CI env for reprotest support

When building .deb packages, we need the python version for the
packaging environment to match that of the target platform, i.e.
python3.7 for buster. In CI, our platform options are:

  * VM, Ubuntu 20.04
  * Container, Debian 10

The container driver in CircleCI does not permit "setarch" calls,
erroring out immediately. The setarch calls are not optional in
reprotest, unfortunately, so let's hack the file and remove it entirely,
only in CI.